refactor: split out scanForDeadObjects to a new file
no behavior changes
